+++ front/app/component-library/components/Shimmer/index.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+
+import { lighten } from 'polished';
+import styled from 'styled-components';
+
+import { colors } from '../../utils/styleUtils';
+import Box, { BoxProps } from '../Box';
+
+const ShimmerBox = styled(Box)<{ $bgColor: string }>`
+  ${({ $bgColor }) => {
+    return `background: linear-gradient(
+      -45deg, 
+      ${$bgColor} 40%, 
+      ${lighten(0.4)($bgColor)} 50%,
+      ${$bgColor} 60%
+    );`;
+  }}
+  background-size: 300%;
+  background-position-x: 100%;
+  animation: shimmer 2s infinite;
+
+  @keyframes shimmer {
+    to {
+      background-position-x: 0%;
+    }
+  }
+`;
+
+const Shimmer = ({ bgColor = colors.black, ...props }: BoxProps) => {
+  return <ShimmerBox $bgColor={bgColor} {...props} />;
+};
+
+export default Shimmer;
